# Optimize token_classification/filter.py and find_label_issues_batched.py for performance